move common parts to base file
