+++ packages/syft/src/syft/service/code/status_sql.py
@@ -13,10 +13,10 @@ import syft as sy
 
 # relative
 from ...types.uid import UID
-from ..job.job_sql import Base
-from ..job.job_sql import CommonMixin
-from ..job.job_sql import PermissionMixin
-from ..job.job_sql import UIDTypeDecorator
+from ..job.base_sql import Base
+from ..job.base_sql import CommonMixin
+from ..job.base_sql import PermissionMixin
+from ..job.base_sql import UIDTypeDecorator
 from .user_code import UserCodeStatusCollection
