+++ packages/syft/src/syft/service/code/user_code_sql.py
@@ -17,10 +17,10 @@ from ...abstract_server import ServerSideType
 from ...server.credentials import SyftVerifyKey
 from ...types.datetime import DateTime
 from ...types.uid import UID
-from ..job.job_sql import Base
-from ..job.job_sql import CommonMixin
-from ..job.job_sql import PermissionMixin
-from ..job.job_sql import UIDTypeDecorator
+from ..job.base_sql import Base
+from ..job.base_sql import CommonMixin
+from ..job.base_sql import PermissionMixin
+from ..job.base_sql import UIDTypeDecorator
 from .user_code import UserCode
 
 if TYPE_CHECKING:
