+++ packages/syft/src/syft/service/output/execution_output_sql.py
@@ -15,10 +15,10 @@ import syft as sy
 # relative
 from ...server.credentials import SyftVerifyKey
 from ...types.uid import UID
-from ..job.job_sql import Base, VerifyKeyTypeDecorator
-from ..job.job_sql import CommonMixin
-from ..job.job_sql import PermissionMixin
-from ..job.job_sql import UIDTypeDecorator
+from ..job.base_sql import Base, VerifyKeyTypeDecorator
+from ..job.base_sql import CommonMixin
+from ..job.base_sql import PermissionMixin
+from ..job.base_sql import UIDTypeDecorator
 
 if TYPE_CHECKING:
     from ..job.job_sql import JobDB
