+++ packages/syft/src/syft/service/project/project_sql.py
@@ -14,10 +14,10 @@ from syft.service.project.project import Project
 # relative
 from ...server.credentials import SyftVerifyKey
 from ...types.uid import UID
-from ..job.job_sql import Base, VerifyKeyTypeDecorator
-from ..job.job_sql import CommonMixin
-from ..job.job_sql import PermissionMixin
-from ..job.job_sql import UIDTypeDecorator
+from ..job.base_sql import Base, VerifyKeyTypeDecorator
+from ..job.base_sql import CommonMixin
+from ..job.base_sql import PermissionMixin
+from ..job.base_sql import UIDTypeDecorator
 
 
 class ProjectDB(CommonMixin, Base, PermissionMixin):
